requirements.txt
pydantic>=2.0
requests>=2.30.0
python-dotenv>=1.0.0
# Add other dependencies as they are identified and used
